mofe offline llm tests into test_llm_with_multi_loras.py
